refactor: use `src` directory
Clears up the service and makes it explicit what is exported.
