+++ source-code/website/src/services/auth/index.server.ts
@@ -1,0 +1,2 @@
+export { router } from "./src/router.js";
+export * from "./src/implementation.js";
