+++ source-code/website/src/services/auth/index.ts
@@ -1,0 +1,2 @@
+export { SignInDialog } from "./src/components/SignInDialog.jsx";
+export { onSignOut } from "./src/onSignOut.js";
