+++ source-code/website/src/services/auth/src/onSignOut.ts
@@ -1,5 +1,5 @@
 import type { SetStoreFunction } from "solid-js/store";
-import type { LocalStorageSchema } from "../local-storage/schema.js";
+import type { LocalStorageSchema } from "../../local-storage/schema.js";
 
 /**
  * This function is called when the user clicks the "Sign Out" button.
